Add files via upload
using PowerShell git commit needed my login credentials but i am unsure of how to do that exactly and i was already cutting it close since i had quite a few problems using the stack with array list to solve the cube. it also said that javac "javac : The term 'javac' is not recognized as the name of a cmdlet, function, script file, or operable program. Check
the spelling of the name, or if a path was included, verify that the path is correct and try again." i could probably figure it out but not while its still today